Bridge/requirements.txt
appnope==0.1.0
attrs==20.2.0
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
base58==2.1.0
bcrypt==3.2.0
bitarray==1.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.1
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
dnspython==1.16.0
ecdsa==0.16.1
et-xmlfile==1.0.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Faker==4.17.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JSON==0.3.4
flask-observability==1.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.0.4
hdwallet==1.1.1
hexbytes==0.2.1
hyperlink==20.0.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.4
incremental==17.5.0
influxdb==5.3.0
iniconfig==1.1.1
ipfshttpclient==0.6.1
ipython==7.18.1
ipython-genutils==0.2.0
itemadapter==0.1.1
itemloaders==1.0.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.1
lru-dict==1.1.7
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
minio==6.0.0
mnemonic==0.19
mongoengine==0.20.0
msgpack==0.6.1
multiaddr==0.0.9
netaddr==0.8.0
networkx==2.5
nltk==3.5
numpy==1.19.4
opencv-python==4.4.0.46
openpyxl==3.0.7
packaging==20.4
pandas==1.1.4
parsel==1.6.0
parsimonious==0.8.1
parso==0.7.1
pdf2image==1.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==3.15.8
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
PyDispatcher==2.0.5
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
pytesseract==0.3.6
pytest==6.1.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2020.4
PyWavelets==1.1.1
queuelib==1.5.0
redis==3.5.3
regex==2020.10.28
requests==2.24.0
rlp==2.0.1
rq==1.5.2
rstr==2.2.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
service-identity==18.1.0
six==1.15.0
sklearn-crfsuite==0.3.6
SQLAlchemy==1.4.9
tabulate==0.8.7
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.11.18
toml==0.10.1
toolz==0.11.1
tqdm==4.53.0
traitlets==5.0.5
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.25.11
varint==1.0.2
w3lib==1.22.0
wcwidth==0.2.5
web3==5.12.2
websockets==8.1
Werkzeug==1.0.1
xlrd==2.0.1
zipp==3.4.0
zope.interface==5.1.2